AfricaEnergy/pyproject.toml
[project]
name = "africaenergy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "python-dotenv",
    "pymongo",
    "selenium",
    "webdriver-manager",
    "beautifulsoup4",
    "certifi",
]
